Reorganize Atom implementations
