+++ src/rules/atoms/mod.rs
@@ -1,0 +1,5 @@
+//! Implementations of `PlanAtom` and `ExecAtom`.
+
+pub mod data;
+pub mod anti;
+pub mod logic;
